Update totalPrice data type OrderItem.cs

+++ Models/OrderItem.cs
@@ -7,8 +7,8 @@ public class OrderItem
     double price { get; set; }
     int quantity { get; set; }
     int idProduct { get; set; }
-    int totalPrice { get; set; }
-  public OrderItem(int idOrderItem, int idOrder, double price, int quantity, int idProduct, int totalPrice)
+    double totalPrice { get; set; }
+  public OrderItem(int idOrderItem, int idOrder, double price, int quantity, int idProduct, double totalPrice)
     {
         this.idOrderItem = idOrderItem;
         this.idOrder = idOrder;
